% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding_functions.R
\name{address2LatLon}
\alias{address2LatLon}
\title{Geocodes data using RDSTK in a loop}
\usage{
address2LatLon(addresses)
}
\arguments{
\item{addresses}{vector of addresses}
}
\value{
dataframe of addresses, lat, and long
}
\description{
This function loads geocodes data using RDSTK, and operates in loops
of chunks set with the chunk index function. It is an intermediary function
used in other geocoding functions in this package
}
