% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty_data_from_census.R
\name{append_poverty_data}
\alias{append_poverty_data}
\title{Appends Poverty Data to main Data file}
\usage{
append_poverty_data(ID_var, pov_column)
}
\arguments{
\item{ID_var}{column in data frame with CT ids, i.e. NEA$CT_GEOID}

\item{pov_column}{the name of the column in the dataset to append
poverty data to}
}
\value{
A vector of poverty data, in the same order as original dataset
}
\description{
This function makes it easy to append 2011-2015 poverty rate data
from the 5 year ACS to another dataset. To use, just create a new column
in the original data to store the poverty rate data, and identify which column
in the original data has the census tract ID variables. The function will do the
rest.
}
\examples{
NEA$poverty_rate<-NA
NEA$poverty_rate<-append_poverty_data(NEA$CT_GEOID, NEA$poverty_rate)
}
