% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geography_based_vars.R
\name{create_boolean_urban}
\alias{create_boolean_urban}
\title{Create Rural/Urban Identifier}
\usage{
create_boolean_urban(MSA_category, Latitude)
}
\arguments{
\item{MSA_category}{Name of column with MSA categories - i.e. M1, M2. should be LSAD}

\item{Latitude}{Column/vector with latitude}
}
\value{
a new column with markers for Rural or Urban
}
\description{
This function can be used to create a column that flags records as either urban
or rural.
}
\details{
Urban records are records that fall into a Metropolitan or Metropolitan NECTA
statistical area. As a result, MSAs will need to be appended to the data before
this function is utilized. Records not falling into these MSAs are classified
as rural. Records without valid Lat/Long will result in NAs

Note, caution should be exercised with international records, they are defaulted
to "rural" since they will not match a MSA. After adding rural/urban
a international record flag should be used to subset the data.
}
