% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NEA_based_vars.R
\name{create_discipline_tag}
\alias{create_discipline_tag}
\title{Create New Discipline Variable}
\usage{
create_discipline_tag(discipline)
}
\arguments{
\item{discipline}{Name of column in data set with "Discipline" variable}
}
\value{
A new column with recoded disciplines for use in analysis
}
\description{
This function is based on the "Discipline" variable in the NEA admin data
and recodes "Discipline" into more succinct categories. The New Discipline
Variable created is used in Tableau visualizations
}
