% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding_functions.R
\name{create_full_address}
\alias{create_full_address}
\title{Combines relevant columns to create a complete address}
\usage{
create_full_address(data, address.vars)
}
\arguments{
\item{data}{name of dataset}

\item{address.vars}{vector of address variables i.e. c("Address1", "Address2", "City")}
}
\value{
formatted vector of addresses
}
\description{
This function combines the address component columns in a dataframe 
and returns an address vector. Also includes functions to remove
problematic characters from the text that will interfere with geocoding
}
\examples{
NEA$full_address<-create_full_address(NEA, c("CoAddress1", "CoAddress2",
"CoCity", "CoState", "CoZip"))
}
