% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intl_flag.R
\name{create_intl_flag}
\alias{create_intl_flag}
\title{Create Flag for International Records}
\usage{
create_intl_flag(state_var, additions = c())
}
\arguments{
\item{state_var}{Name of column in data set with state abbreviations}

\item{additions}{Additional State Names to add as international.
defaults to "FO", "AS", "FM", "GU","MH", "MP", "PW", "PR", "VI",
 "AE","AP", "AA", "CM"}
}
\value{
A new Boolean TRUE and FALSE Column where TRUE=International
}
\description{
This function can be used to create a column that flags intl records. It assumes 
that the dataset includes a column with State Names
}
