% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geography_based_vars.R
\name{create_poverty_flag}
\alias{create_poverty_flag}
\title{Create Poverty/Not Poverty Flag}
\usage{
create_poverty_flag(pov_column, Latitude, cutoff = 20)
}
\arguments{
\item{pov_column}{Name of column with poverty rate - i.e. 20, 30 etc.}

\item{Latitude}{Column/vector with latitude}

\item{cutoff}{cutoff for defining a high poverty neighborhood, defaults to 20}
}
\value{
a new column with markers for Rural or Urban
}
\description{
This function can be used to create a column that categorizes
records based on their poverty rate.
}
\details{
It requires that the dataset already has a column for poverty rate
for the census tract in which the record is located, 
which can be created using the relevant poverty packages in this
r package.

Note, that areas without population or without poverty rates are labeled
"Missing Data"
}
