% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geography_based_vars.R
\name{create_urban_type}
\alias{create_urban_type}
\title{Create Categories of Urban Records}
\usage{
create_urban_type(MSA_pop, Boolean_Urban_MSA)
}
\arguments{
\item{MSA_pop}{Column with population of the MSA. Make sure to manually add in NECTA pop as well.}

\item{Boolean_Urban_MSA}{column with rural/urban marker}
}
\value{
a new column with categories for type of urban area
}
\description{
This function can be used to create a column that categorizes
records based on urban population size
}
\details{
It requires that the dataset already has a rural/urban marker, and that
it has a second column with MSA and NECTA population data - should
be downloaded and appended from the Census, and matched based on CBSA and 
NECTA ids.

Note, caution should be exercised with international records, they are defaulted
to "rural" since they will not match a MSA. After adding rural/urban
a international record flag should be used to subset the data.
}
