% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_matching_functions.R
\name{get_cd_data}
\alias{get_cd_data}
\title{Get Dataframe of Matching Congressional District}
\usage{
get_cd_data(Latitude, Longitude)
}
\arguments{
\item{Latitude}{vector of Latitudes}

\item{Longitude}{Vector of Longitudes}
}
\value{
A datafame with associated CD info. Will be in correct order
as original Lat/Long to make it easy to append the data
}
\description{
This function returns a dataframe with Congressional data for a vector of Latitudes
and Longitudes in a dataframe.
}
\details{
The package operates by matching Lat/Long to shapefiles of 114th Congressional
Districts - valid from 2015-2017.

It returns a dataframe with the following components:
 GEOID: The GEOID of the CD
 NAME_LSAD: The Name of the CD
}
\examples{
cds<-get_cd_data(NEA$CoLatitude, NEA$CoLongitude)
NEA$CD_GEOID<-cds$CD_GEOID
NEA$CD_NameLSAD<-cds$CD_NameLSAD
}
