% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_matching_functions.R
\name{get_county_data}
\alias{get_county_data}
\title{Get Dataframe of Matching County Data}
\usage{
get_county_data(Latitude, Longitude, year = 2015)
}
\arguments{
\item{Latitude}{vector of Latitudes}

\item{Longitude}{Vector of Longitudes}

\item{year}{vintage of shapefiles, defaults to 2015 if none specified}
}
\value{
A datafame with associated Ccountyinfo. Will be in correct order
as original Lat/Long to make it easy to append the data
}
\description{
This function returns a dataframe with County ID data for a vector of Latitudes
and Longitudes in a dataframe.
}
\details{
It returns a dataframe with the following components:
 GEOID: The GEOID of the County - i.e. FIPS Code
 NAME_LSAD: The Name of the County
}
\examples{
counties<-get_cunty_data(NEA$CoLatitude, NEA$CoLongitude)
NEA$county_GEOID<-counties$county_GEOID
NEA$county_NameLSAD<-counties$county_NameLSAD
}
