% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding_functions.R
\name{get_geocode_data}
\alias{get_geocode_data}
\title{Primary Geocoding Function}
\usage{
get_geocode_data(address)
}
\arguments{
\item{address}{name of the address vector in the data set i.e. NEA$full_address}
}
\value{
dataframe of addresses and geocodes, that can then be added to the original
data
}
\description{
This function takes an address column that has been properly formatted
using get_padded_zip and then create_full_address and returns a dataframe of 
addresses and relevant lat/long
}
\examples{
addresses<-get_geocode_data(NEA$full_address)
NEA$CoLongitude<-addresses$long
NEA$CoLatitude<-addresses$lat
}
