% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding_functions.R
\name{get_padded_zip}
\alias{get_padded_zip}
\title{Pads zipcodes to 5 digits}
\usage{
get_padded_zip(zip)
}
\arguments{
\item{zip}{vector of zipcodes, five digit format}
}
\value{
vector of padded zipcodes
}
\description{
This function uses the stringr package to automate
padding zipcodes to 5 digits
}
\examples{
NEA$zip<-get_padded_zip(NEA$CoZip)
}
