% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty_data_from_census.R
\name{get_poverty_rates}
\alias{get_poverty_rates}
\title{Creates a poverty rate variable, and merges two files
with different census tables}
\usage{
get_poverty_rates(poverty, population)
}
\arguments{
\item{poverty}{dataframe with poverty table}

\item{population}{dataframe with population table}
}
\value{
A dataframe with formatted/merged pop and poverty data
}
\description{
This function requires two datasets using the tract level data function:
-a dataframe that has the poverty table data  "B17021_002E"
-a second dataframe that has the population data for which poverty is 
known "B17021_001E"
}
