% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty_data_from_census.R
\name{tract_level_data}
\alias{tract_level_data}
\title{Download CT Level data from Census API}
\usage{
tract_level_data(year = "2015", survey = "acs5", table = "B17021_002E")
}
\arguments{
\item{year}{ending year of ACS data, defaults to 2015}

\item{survey}{survey used, defaults to acs5}

\item{table}{table used, defaults to "B17021_002E"}
}
\value{
A dataframe with census data
}
\description{
This function downloads a national-level file of tract-level data
from a specified census table. Data is downloaded in a loop
}
